@@ -1,176 +1,113 @@
 package com.rpncalc;
 
-import com.rpncalc.exception.BusinessException;
+import com.rpncalc.exception.CalculatorException;
 import com.rpncalc.exception.ErrorCodeEnum;
-import com.rpncalc.operator.function.ClearFunctionOperator;
-import com.rpncalc.operator.Operator;
-import com.rpncalc.operator.function.UndoFunctionOperator;
-import com.rpncalc.operator.math.AdditionOperator;
-import com.rpncalc.operator.math.DivisionOperator;
-import com.rpncalc.operator.math.MultiplicationOperator;
-import com.rpncalc.operator.OperatorTypeEnum;
-import com.rpncalc.operator.math.SqrtOperator;
-import com.rpncalc.operator.math.SubtractionOperator;
-import com.rpncalc.stack.OperatorValueStack;
-import org.apache.commons.lang3.ArrayUtils;
-import org.apache.commons.lang3.StringUtils;
+import com.rpncalc.operator.Operation;
+import com.rpncalc.operator.OperationFactory;
+import com.rpncalc.operator.function.UndoOperation;
+import com.rpncalc.snapshot.Caretaker;
+import com.rpncalc.snapshot.Memento;
+import com.rpncalc.stack.OperandStack;
+import com.rpncalc.stack.OperandStackImpl;
+import com.rpncalc.stack.OperandValue;
+
 import java.math.BigDecimal;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.Arrays;
 
 public class Calculator {
 
-    private OperatorValueStack stack = new OperatorValueStack();
+    private OperandStack operandStack;
 
-    private Map<String, Operator> operatorMap;
+    private Caretaker caretaker;
 
-    private String[] tokens;
-    private int tokenCursor;
-    private String express;
+    private OperationFactory operationFactory;
 
     public Calculator() {
-        tokens = new String[0];
-        tokenCursor = 0;
-        express = "";
-
-        initOperatorMap();
+        this.operandStack = new OperandStackImpl();
+        this.caretaker = new Caretaker();
+        this.operationFactory = new OperationFactory(operandStack, caretaker);
     }
 
-    private void initOperatorMap() {
-        operatorMap = new HashMap<>();
+    private void processOperator(Operation operation) throws CalculatorException {
 
-        Operator[] ops = new Operator[]{
-                new AdditionOperator(),
-                new SubtractionOperator(),
-                new MultiplicationOperator(),
-                new DivisionOperator(),
-                new SqrtOperator(),
-                new ClearFunctionOperator(stack),
-                new UndoFunctionOperator(stack),
-        };
+        int operandNum = operation.operandNum();
 
-        for (Operator op : ops) {
-            operatorMap.put(op.symbol(), op);
-        }
-    }
-
-    private int getOriginalTokenPos() {
-
-        int tokenIndex = 0;
-        int pos = 0;
-
-        while (pos < express.length()) {
-
-            if (express.charAt(pos) != ' ') {
-                if (tokenIndex == tokenCursor) {
-                    return pos + 1;
-                }
-                ++tokenIndex;
-                while (pos < express.length() && express.charAt(pos) != ' '){
-                    ++pos;
-                }
-            }
-            ++pos;
+        if (operandNum == 0) {
+            // Execute a function operation
+            operation.doFunction();
         }
 
-        return -1;
-    }
+        else {
+            // Execute a math operation
+            BigDecimal[] operands = Arrays
+                    .stream(operandStack.pop(operandNum))
+                    .map(OperandValue::getValue).toArray(BigDecimal[]::new);
 
+            BigDecimal resultVal = operation.calculate(operands);
 
-    private void processOperator(Operator operator, OperatorTypeEnum type) throws BusinessException {
-
-        int paramNum = type.getParamNum();
-        if (stack.size() < paramNum) {
-            String message = String.format(
-                    ErrorCodeEnum.INSUFFICIENT_PARAMETERS.getMessage(), tokens[tokenCursor], getOriginalTokenPos());
-
-            throw new BusinessException(ErrorCodeEnum.INSUFFICIENT_PARAMETERS, message);
-        }
-
-        switch (paramNum) {
-            case 0:
-                operator.execute();
-                break;
-
-            case 1:
-                BigDecimal input = stack.pop();
-                stack.push(operator.execute(input));
-                break;
-
-            case 2:
-                stack.startBatch();
-
-                // The top element is right
-                BigDecimal right = stack.pop();
-                BigDecimal left = stack.pop();
-
-                stack.push(operator.execute(left, right));
-                stack.endBatch();
-                break;
-
-            default:
-                throw new BusinessException(
-                        ErrorCodeEnum.INTERNAL_ERROR, "Unsupported operator:" + type.getSymbol());
+            operandStack.push(new OperandValue(resultVal));
         }
     }
 
-    private void processToken(String token) throws BusinessException {
-
-        if (operatorMap.containsKey(token)) {
-            Operator operator = operatorMap.get(token);
-            OperatorTypeEnum type = OperatorTypeEnum.fromSymbol(token);
-            processOperator(operator, type);
-            return;
-        }
-
+    private void processNumber(String token) throws CalculatorException {
         try {
-            stack.push(new BigDecimal(token, Constant.DECIMAL_STORE_PRECISION));
+            BigDecimal value = new BigDecimal(token);
+            operandStack.push(new OperandValue(value));
 
         } catch (NumberFormatException e) {
-            throw new BusinessException(
-                    ErrorCodeEnum.INVALID_TOKEN,
-                    String.format(ErrorCodeEnum.INVALID_TOKEN.getMessage(), token, getOriginalTokenPos()));
+            throw new CalculatorException(ErrorCodeEnum.INVALID_TOKEN);
         }
     }
 
-    /**
-     * This function is used to display the remained tokens that are ignored because of previous token error.
-     *
-     * @return The remained tokens that were cut off
-     */
-    public String[] getRemainedTokens() {
-        if (tokenCursor != tokens.length) {
-            return ArrayUtils.subarray(tokens, tokenCursor + 1, tokens.length);
+    private void processToken(String token) throws CalculatorException {
+
+        Operation operation = operationFactory.create(token);
+        if (operation != null) {
+            processOperator(operation);
+
+        } else {
+            processNumber(token);
         }
 
-        return new String[0];
+        // Take snapshot
+        if (!UndoOperation.TOKEN.equals(token)) {
+            // undo token doesn't need to take snapshot
+            Memento memento = operandStack.snapshot();
+            caretaker.add(memento);
+        }
     }
 
     /**
      * Execute the input RPN expression
      *
-     * @param expression RPN expression
+     * @param rpnExpression RPN expression
      * @return object contains result and error message.
      */
-    public ExecuteResult execute(String expression) {
-        if (StringUtils.isBlank(expression)) {
-            return new ExecuteResult(this);
-        }
+    public ExecuteResult execute(String rpnExpression) {
 
-        tokens = expression.trim().split("\\s+");
-        tokenCursor = 0;
-        express = expression;
+        RpnExpression expression = new RpnExpression(rpnExpression);
+        String[] tokens = expression.getTokens();
 
-        try {
-            for (String token : tokens) {
+        ExecuteResult result = new ExecuteResult(expression);
+
+        for(int pos = 0; pos < tokens.length; ++pos) {
+
+            try {
+                String token = tokens[pos];
+                result.setToken(token, pos);
+
                 processToken(token);
-                ++tokenCursor;
-            }
 
-        } catch (BusinessException e) {
-            return new ExecuteResult(this, e.getMessage(), e.getCode(), stack.toString());
+            } catch (CalculatorException e) {
+                result.setError(e.getCode());
+                return result;
+
+            } finally {
+                result.setResult(operandStack.toString());
+            }
         }
 
-        return new ExecuteResult(this, stack.toString());
+        return result;
     }
+
 }
