@@ -2,41 +2,53 @@ package com.rpncalc.exception;
 
 import lombok.AllArgsConstructor;
 
+import java.util.Arrays;
+
 @AllArgsConstructor
 public enum ErrorCodeEnum {
 
     /**
      * Success
      */
-    SUCCESS("000000", "Success"),
+    SUCCESS(0, "Success"),
+
+    /**
+     * UNKNOWN ERROR
+     */
+    UNKNOWN_ERROR(-1, "Unknown error"),
 
     /**
      * INTERNAL ERROR
      */
-    INTERNAL_ERROR("000200", "Internal error"),
+    INTERNAL_ERROR(-2, "Internal error"),
 
     /**
      * Arithmetic Errors
      */
-    DIVISION_BY_ZERO("000300", "Division by zero"),
-    NEGATIVE_SQUARE_ROOT("000300", "Square root is negative"),
+    DIVISION_BY_ZERO(-3, "Division by zero"),
+    NEGATIVE_SQUARE_ROOT(-4, "Square root is negative"),
 
     /**
      * Invalid Expression
      */
-    INSUFFICIENT_PARAMETERS("000400", "operator %s (position: %d): insufficient parameters"),
-    INVALID_TOKEN("000401", "token %s (position: %d): invalid token"),
+    INSUFFICIENT_PARAMETERS(-5, "operator %s (position: %d): insufficient parameters"),
+    INVALID_TOKEN(-6, "token %s (position: %d): invalid token"),
     ;
 
-    private String code;
+    private Integer code;
     private String message;
 
-    public String getCode() {
+    public Integer getCode() {
         return code;
     }
 
     public String getMessage() {
         return message;
+    }
+
+    public static ErrorCodeEnum valueOf(int code) {
+        ErrorCodeEnum[] es = ErrorCodeEnum.values();
+        return Arrays.stream(es).filter(e -> e.code == code).findAny().orElse(UNKNOWN_ERROR);
     }
 }
 
